backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
browser-use>=0.12.0
pydantic>=2.0.0
